Factor out transactions card from block details
