+++ src/app/pages/BlockDetailPage/index.tsx
@@ -1,9 +1,7 @@
 import { FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useParams } from 'react-router-dom'
-import Card from '@mui/material/Card'
-import CardHeader from '@mui/material/CardHeader'
-import CardContent from '@mui/material/CardContent'
+
 import Skeleton from '@mui/material/Skeleton'
 import useMediaQuery from '@mui/material/useMediaQuery'
 import { useTheme } from '@mui/material/styles'
@@ -12,11 +10,9 @@ import { StyledDescriptionList } from '../../components/StyledDescriptionList'
 import { PageLayout } from '../../components/PageLayout'
 import { SubPageCard } from '../../components/SubPageCard'
 import { CopyToClipboard } from '../../components/CopyToClipboard'
-import { Transactions } from '../../components/Transactions'
-import { useSearchParamsPagination } from '../../components/Table/useSearchParamsPagination'
-import { NUMBER_OF_ITEMS_ON_SEPARATE_PAGE } from '../../config'
-import { useGetEmeraldTransactions } from '../../../oasis-indexer/api'
+
 import { useFormattedTimestampString } from '../../hooks/useFormattedTimestamp'
+import { TransactionsCard } from './TransactionsCard'
 
 // TODO: replace with an appropriate API
 function useGetEmeraldBlockByHeight(blockHeight: number) {
@@ -29,35 +25,14 @@ function useGetEmeraldBlockByHeight(blockHeight: number) {
 }
 
 export const BlockDetailPage: FC = () => {
-  const { t } = useTranslation()
   const blockHeight = parseInt(useParams().blockHeight!, 10)
   const { isLoading, data } = useGetEmeraldBlockByHeight(blockHeight)
   const block = data.data
-  const txsPagination = useSearchParamsPagination('page')
-  const txsOffset = (txsPagination.selectedPage - 1) * NUMBER_OF_ITEMS_ON_SEPARATE_PAGE
-  const transactionsQuery = useGetEmeraldTransactions({
-    block: blockHeight,
-    limit: NUMBER_OF_ITEMS_ON_SEPARATE_PAGE,
-    offset: txsOffset,
-  })
 
   return (
     <PageLayout>
       <BlockDetailView isLoading={isLoading} block={block}></BlockDetailView>
-      <Card>
-        <CardHeader disableTypography component="h3" title={t('common.transactions')} />
-        <CardContent>
-          <Transactions
-            transactions={transactionsQuery.data?.data.transactions}
-            isLoading={transactionsQuery.isLoading}
-            limit={NUMBER_OF_ITEMS_ON_SEPARATE_PAGE}
-            pagination={{
-              selectedPage: txsPagination.selectedPage,
-              linkToPage: txsPagination.linkToPage,
-            }}
-          />
-        </CardContent>
-      </Card>
+      <TransactionsCard blockHeight={blockHeight} />
     </PageLayout>
   )
 }
